+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/azure
 
+## 1.3.3
+
+### Patch Changes
+
+- Updated dependencies [24befd8]
+  - @ai-sdk/openai@1.3.3
+
+## 1.3.2
+
+### Patch Changes
+
+- Updated dependencies [db15028]
+  - @ai-sdk/openai@1.3.2
+
 ## 1.3.1
 
 ### Patch Changes