+++ packages/solid/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/solid
 
+## 1.2.3
+
+### Patch Changes
+
+- d92fa29: feat: add credentials support to experimental useObject and StructuredObject
+
+## 1.2.2
+
+### Patch Changes
+
+- 5185bce: chore (ui/solid): deprecate ai-sdk/solid
+
 ## 1.2.1
 
 ### Patch Changes